add bullet rmv on collision with map tile

+++ src/models/CollisionModel.java
@@ -15,10 +15,10 @@ public class CollisionModel {
     public CollisionModel(Vector2f position, int model_width, int model_height) {
         this.position = position;
         this.corners = new Point[4];
-        this.corners[0] = new Point(-model_width / 2, -model_height / 2);    // TOP LEFT
-        this.corners[1] = new Point(model_width / 2, -model_height / 2);     // TOP RIGHT
-        this.corners[2] = new Point(model_width / 2, model_height / 2);      // BOTTOM RIGHT
-        this.corners[3] = new Point(-model_width / 2, model_height / 2);     // BOTTOM LEFT
+        this.corners[0] = new Point((float) -model_width / 2.f, (float) -model_height / 2.f);    // TOP LEFT
+        this.corners[1] = new Point((float) model_width / 2.f, (float) -model_height / 2.f);     // TOP RIGHT
+        this.corners[2] = new Point((float) model_width / 2.f, (float) model_height / 2.f);      // BOTTOM RIGHT
+        this.corners[3] = new Point((float) -model_width / 2.f, (float) model_height / 2.f);     // BOTTOM LEFT
         this.collision_points = new Point[4];
         collision_points[0] = new Point(0, 0);
         collision_points[1] = new Point(0, 0);
